+++ src/model/Product.java
@@ -22,13 +22,15 @@ package model;////**Wymagania Biznesowe**
 
 //DONE
 
+import java.math.BigDecimal;
+
 public class Product {
     private int id;
     private String name;
-    private double price;
+    private BigDecimal price;
     private int availableQuantity;
 
-    public Product(int id, String name, double price, int availableQuantity) {
+    public Product(int id, String name, BigDecimal price, int availableQuantity) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -51,11 +53,11 @@ public class Product {
         this.name = name;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
